Refactor animation service to handle path and nodeOrder as Node arrays
- Updated animationService to directly assign path and nodeOrder from response without JSON parsing.
- Modified PostTraversalResponse interface to reflect path and nodeOrder as Node arrays instead of strings.

Add GitHub Actions workflow for backend CI

- Created backend continuous integration workflow to run tests and manage dependencies.
- Added steps for setting up Python, installing dependencies, running tests, and deploying documentation to GitHub Pages.
- Included a job for Dependabot to enable auto-merge for its pull requests.

Add pre-commit configuration for code quality checks

- Introduced a pre-commit configuration file to enforce coding standards and run various checks on commits.
- Configured hooks for Python type checking, formatting, and validation of markdown and JSON files.

+++ fast-api/src/map.py
@@ -23,7 +23,7 @@ class Node:
     """
 
     position: Position
-    weight: float
+    weight: int
 
     def to_dict(self):
         return {"x": self.position.x, "y": self.position.y, "weight": self.weight}
